requirements.txt
streamlit
python-dotenv
langchain-community
langchain-openai
langchain